Use a cache of ListOptionValues to solve some problems
As discussed in
https://github.com/rpm-software-management/dnf/pull/1226 ,
@dmach 's initial implementation of this attempt to wrap libdnf
OptionStringLists so they behave more like real Python lists
had a few issues. Taking multiple refs to the same config option
gave you a new ListOptionValue each time, and these could easily
get out of sync with each other, such that changing one then
changing the other would have unexpected results. Also, if you
take a reference to a config option and then *reassign that
option*, like so:

ref = conf.pluginpath
conf.pluginpath = ['/foo']

then subsequent changes to `ref` should no longer affect the
value of `conf.pluginpath`, but in fact they can. This commit
uses a cache of ListOptionValues, stored on the config instance,
to try and solve both problems. The first time a reference is
taken to the config option, we create a new ListOptionValue and
stash it in the cache. Each subsequent time a reference is
taken, we return the same LOV that's currently in the cache. So
when you take multiple refs you really get references to the
same object (just as you do with normal Python lists), and they
cannot get out of sync.

Additionally, we make ListOptionValue instances check whether
they are the same as the instance currently in the cache before
modifying the underlying libdnf config value, and have setattr
put a new LOV in the cache when it's called. This solves the
problem of 'stale' references still being able to change the
underlying value after the config attribute for the option has
been reassigned. With this change 'stale' references behave just
as they did under pure Python - they keep working as Python
lists, but no change to them actually changes the underlying
config value.

Signed-off-by: Adam Williamson <awilliam@redhat.com>