Update ListOptionValue from backing value on access
This feels like it's getting to be a bit crazy, but this at
least seems to really work in all the scenarios I can come up
with. This tries to solve the problem of ListOpenValues
getting out of sync when the underlying config option value
is changed behind their back - e.g. by a call to Option._set()
or BaseConfig._set_value(). This can cause an LOV instance to
no longer show the correct value when queried, and to overwrite
the change if mutated.

To try and deal with that, we extend the wrapping to nearly
every method - all the ones that actually involve the list's
values, so far as I know. We have the wrapper update the value
from the backing store before it does much else, *so long as*
the LOV is still the 'canonical' one in the cache (if it's been
'decoupled' we do not want it to reflect the backing store any
more, so we don't) *and* it is not a nested call. I found that
if you just make *every* wrapped call do this, some mutations
just don't work, because e.g. doing an `append()` involves a
bunch of calls to `__len__` and `__getitem__`, and it seems that
if *those* calls read in the old value from the backing store,
the value you were trying to append gets lost. So we keep a
counter of wrapper calls that is incremented at the start of
the wrapper function and decremented at the end, and only update
the value from the backing store when the counter value is 1.

Signed-off-by: Adam Williamson <awilliam@redhat.com>